requirements_pytest.txt
pypdf==5.1.0
requests==2.32.3
moviepy==2.1.1
numpy==2.2.1
pillow<11.0.0
pytest-playwright==0.6.2
asyncio==3.4.3
pytest-asyncio==0.25.0
pyperclip==1.9.0
pyotp==2.9.0
reportlab==4.2.5